+++ src/brush/brush.h
@@ -14,7 +14,7 @@ namespace sf {
 }
 
 struct BrushSettings {
-    int size = 5;
+    float size = 5;
     int feathering = 0;
     sf::Color color = sf::Color::White;
 
@@ -26,9 +26,9 @@ public:
     Brush() = default;
     explicit Brush(Canvas& canvas);
     virtual ~Brush() = default;
-    virtual void onMouseDown(const sf::Vector2i& pos) = 0;
-    virtual void onMouseUp(const sf::Vector2i& pos) = 0;
-    virtual void onMouseMoved(const sf::Vector2i& pos) = 0;
+    virtual void onMouseDown(const sf::Vector2f& pos) = 0;
+    virtual void onMouseUp(const sf::Vector2f& pos) = 0;
+    virtual void onMouseMoved(const sf::Vector2f& pos) = 0;
     void setCanvas(Canvas& canvas);
     inline static BrushSettings settings{};
 
@@ -37,12 +37,6 @@ public:
     static void initShader();
 
 protected:
-//    void circle(const sf::Vector2i& pos);
-//    void draw_line(const sf::Vector2i& pos1, const sf::Vector2i& pos2, int a, int b, int c);
-//    void line(const sf::Vector2i& pos1, const sf::Vector2i& pos2);
-//
-//    void setPixel(const sf::Vector2i& pos, const sf::Color& color);
-
     Canvas* _canvas{};
     static sf::Shader renderShader;
 };
